week_3/requirements.txt
terminaltables==3.1.0
tensorflow_gpu==1.13.1
imgaug==0.3.0
scipy==1.4.1
six==1.12.0
filterpy==1.4.5
torchvision==0.3.0
motmetrics==1.2.0
torch==1.1.0
numpy==1.16.4
pip==20.0.2
scikit_image==0.15.0
Keras==2.2.4
pytest==5.3.5
~pencv_python==4.1.0.25
pycocotools==2.0
h5py==2.9.0
tqdm==4.41.1
imageio==2.5.0
keras_resnet==0.1.0
matplotlib==3.1.3
classification_models==0.1
efficientnet==1.1.0
ipython==7.13.0
Pillow==7.0.0
progressbar33==2.4
scikit_learn==0.22.2.post1
skimage==0.0
tensorflow==2.1.0
